requirements.txt
Django==1.7.5
django-cors-headers==1.0.0
django-uuidfield==0.5.0
djangorestframework==3.1.0
Pillow==2.7.0
